Leaving Rail Connections for Now
The rail connecting is being a bit iffy, so I'm going to leave it for now. The car can go from one rail to another, but looping back to the start is iffy as well. If the car reaches the end and isn't moving setting it to reverse won't do anything.