Starting with some UI elements.
Starting with some UI elements. You're going to need a button for going back to the main menu. I think what you'll do is manually create the tower, and then have the game start determine which tower to use. So this will require you to determine what rail to enable, and what rail to disable. Also should probably figure out a system  to respawn blocks that fall out of the play area.